--- a/DataFormats/EgammaCandidates/interface/GsfElectron.h
+++ b/DataFormats/EgammaCandidates/interface/GsfElectron.h
@@ -183,10 +183,6 @@ class GsfElectron : public RecoCandidate
     // forward core methods
     SuperClusterRef superCluster() const override { return core()->superCluster() ; }
     GsfTrackRef gsfTrack() const override { return core()->gsfTrack() ; }
-    //so it turns out closestCtfTrackRef is overriden by pat::Electron but not closestTrack()
-    //this results in surprising behaviour, ie two functions which looked identical 
-    //would give different results so now closestTrack just calls closestCtfTrackRef
-    virtual TrackRef closestTrack() const { return closestCtfTrackRef() ;}
     float ctfGsfOverlap() const { return core()->ctfGsfOverlap() ; }
     bool ecalDrivenSeed() const { return core()->ecalDrivenSeed() ; }
     bool trackerDrivenSeed() const { return core()->trackerDrivenSeed() ; }