--- a/DataFormats/EgammaCandidates/src/GsfElectron.cc
+++ b/DataFormats/EgammaCandidates/src/GsfElectron.cc
@@ -245,3 +245,12 @@ float GsfElectron::p4Error( P4Kind kind ) const
    }
  }
 
+float GsfElectron::closestTrackChi2() const
+ { 
+   return  closestTrackValid() ? closestTrack()->normalizedChi2() : 0. ; 
+ }
+ 
+int GsfElectron::closestTrackNHits() const
+ { 
+   return closestTrackValid() ? closestTrack()->hitPattern().trackerLayersWithMeasurement() : -1 ;  
+ }