… DataFormats/EgammaCandidates/src/classes_def.xml
… PhysicsTools/NanoAOD/python/photons_cff.py
… PhysicsTools/PatAlgos/python/slimming/miniAOD_tools.py
… RecoEcal/EgammaCoreTools/interface/EcalClusterTools.h
… RecoEgamma/EgammaElectronAlgos/interface/GsfElectronAlgo.h
… RecoEgamma/EgammaElectronProducers/plugins/GsfElectronBaseProducer.cc
… RecoEgamma/EgammaElectronProducers/python/gedGsfElectrons_cfi.py
… RecoEgamma/EgammaElectronProducers/python/gsfElectrons_cfi.py
… RecoEgamma/EgammaElectronProducers/python/lowPtGsfElectrons_cfi.py
… RecoEgamma/EgammaPhotonProducers/interface/GEDPhotonProducer.h
… RecoEgamma/EgammaPhotonProducers/python/gedPhotonSequence_cff.py
… RecoEgamma/EgammaPhotonProducers/python/ootPhotons_cff.py
… RecoEgamma/EgammaPhotonProducers/src/GEDPhotonProducer.cc
… RecoEgamma/EgammaTools/interface/ConversionTools.h
… RecoEgamma/EgammaTools/interface/EleEnergyRetriever.h
… RecoEgamma/EgammaTools/plugins/EG9X105XObjectUpdateModifier.cc
… RecoEgamma/EgammaTools/plugins/EGPfIsolationModifier.cc
… RecoEgamma/EgammaTools/src/ConversionTools.cc
… RecoEgamma/EgammaTools/src/MVAVariableHelper.cc
… RecoEgamma/ElectronIdentification/data/ElectronIDVariables.txt
… RecoEgamma/ElectronIdentification/data/ElectronMVAEstimatorRun2Fall17V1Variables.txt
… RecoEgamma/ElectronIdentification/data/ElectronMVAEstimatorRun2Variables.txt
… RecoEgamma/ElectronIdentification/interface/EBEECutValues.h
… RecoEgamma/ElectronIdentification/interface/ElectronMVAEstimatorRun2.h
… RecoEgamma/ElectronIdentification/plugins/ElectronHEEPIDValueMapProducer.cc
… RecoEgamma/ElectronIdentification/plugins/cuts/GsfEleEmHadD1IsoRhoCut.cc
… RecoEgamma/ElectronIdentification/plugins/cuts/GsfEleMinEcalEtCut.cc
… RecoEgamma/ElectronIdentification/plugins/cuts/GsfEleTrkPtIsoCut.cc
… RecoEgamma/ElectronIdentification/python/Identification/heepElectronID_HEEPV71_cff.py
… RecoEgamma/ElectronIdentification/python/Identification/heepElectronID_tools.py