~~This is a WIP, there are a few outstanding bugs I'm trying to pin down.  I could also split this PR into two, most of the changes is adapting VID to read from the object.  It will not pass unit tests (because of tau problem which generical to all cms) nor matrix (one test still to fix).~~: Now fixed

## PR description:

In E/gamma IDs, variables are sometimes calculated on “on the fly” during the computation of the ID. While we have survived with this, it would be much nicer if the variables were in the electron/photon as this means we dont have to continuously re-run expensive producers. 

This PR does two things: it moves wherever possible an ID variable to the the photon/electron object and ensures the correct ID information is propagated to the miniAOD. At the same time it adds in HEEPV7.1 ID which is a bug fix ID which changes the pt cut to 5 GeV and uses calo energy to scale the isolation cut. 

A mechanism to update <105X samples exists via RecoEgamma/EgammaTools/plugins/EG9X105XObjectUpdateModifier.cc

### Photon Variable changes

#### particle flow isolation:

it turns out only pfcluster isolation, charged, neutral and photon were being filled so the others were removed 

photon and neutral isolation use the isolation calculated by PhotonIDValueMap producer as that is what is used in E/gamma cut based and MVA IDs. These two variables are identifical to the existing photon & neutral isolations.  

charged isolation: this is a practical but not fundamental definition change. The existing charged isolation uses the particle flow vtx association to determine vertex. The current charged isolation uses dxy, dz cuts instead. While the two isolations will disagree, they are fundamentally the same thing so the new variable can replace the existing one. Again this is the variable photon IDs actually use.  The old charged isolation is now chargedHadronPFPVIso incase anybody really wants it. 

chargedHadronWorstVtxIso and chargedHadronWorstVtxGeomVetoIso have been added

#### Other ids: showershape:
smajor, sminor, smajorAlpha used in long lived analyses have been added.  There is a bug long standing bug in the code to calculate them, it doesnt check it actually finds a rec-hit and blindly uses it anyways. Patch incoming for that. 

### Electron Variable changes:

the HEEP track isolation is added to the object at the AOD level (it was already in the miniAOD). I would have liked to simply replace the tracker isolation variable but it changes the PF event content (its used in the ID). There is no reason to suspect that the changes are worse (or better) but at this point its best not to touch it. 

conversion vtx probablity: this is now added to the electron as its used in the MVA. This causing me major problems trying to get it to agree, this variable is very buggy and all the more reason to be in the object!

helperfuncs:  closestTrkNHits() and closetTrkChi2() allows these functions to be more easily used in our framework

The change to the dataformats is just a few files, most of the changes come from VID and adapting it to use the new format. 

### Expected physics changes:
heep trk isolation (and thus ID) will slight change on some events, this is because we’ve gone from calculating on PackedCandidates to actual tracks so we encounter some small differences due to precision. 
photon : photon & neutral hadron isolations will in rare cases change slightly, due to a small difference in vetos between CITK and PhotonID isolation computations. The variables remain fundementally the same though
photon : charged isolation is now a different but conceptually similar variable, the old one remains
IDs : no ID or MVA variable changes should happen except  a few events for the HEEP IDs changing due to rounding errors. 


## PR validation:

An event by event comparison on 2018 EGamma data has been performed. Its not quite closing yet but soon (mainly as I'm finding lots of existing bugs...)

## To Do:

1. ~~problems with calculating the conversion vertex fit probability and some disagreements in the photon MVA.~~
1. ~~3 problems in matrix tests, I know how to fix two of them (its a bug elsewhere), the 3rd I need to resolve~~
1. ~~photon MVA isnt agreement well, something is bugged there~~ now very minor