@Sam-Harper 2019-04-08
Yep as I mentioned, I know the solution for two of those crashes which is improper validity checking of hits found in the cluster2ndMoments. The last (the HION), I'll figure out tomorrow. Additionally I've figured out my discrepancies in the conversion fit probability (its a nasty bug which has been around forever) and the photon MVA problem is just the CITK vs PhotonIDValueMap producer defination of photon / neutral isolation (which should be the same but are not). I'm going to spend an hour or so debugging the photon / neutral isolation bug tomorrow to see what is going on. 

The fall back practical solution is that for the conversion fit vtx prob and photon pf isolation values, I'm going to make a "bugged" data member which stores the delta of the bugged value (which is in use!) with the correct value. It'll be zero most of the time so will compress well.