@Sam-Harper 2019-04-09
so I have fixed my issues. I still have a rare bug giving not identical photon MVA scores but its close enough. 

I also include a fix for the EcalClusterTools::cluster2ndMomentum calculation as it was not checking that it actually found the hit and also removed the taus test which no longer works. 

The conversion fit probablity was found to be a bug in E/gamma ConversionTools (and its semi serious). The issue is that it matches by reference to the conversion tracks (ie product key & index). The probably is that reducedEgamma puts the GsfTracks into a new collection which means the electrons ref is now out of date and therefore the GsfTrack is excluded from being able to match a conversion. This is now sadly baked into E/gamma IDs so thats the new defination.

I'm debating putting a variable (probably as the delta) for proper conversion match.