@Sam-Harper 2019-04-09
now I need to handle the nanoAOD reading different releases however there is another issue https://github.com/cms-nanoAOD/cmssw/pull/308#issuecomment-481187073 which also needs to be resolved and they will happen at the same time