@guitargeek 2019-04-10
Hi Sam, nice PR and thank you also for the summary talk this morning! Did you already study how much the conversion bug affects the ID? It's true that it's a priori safer to stick with the definitions the ID was trained with, but maybe we should at least check? Because if it would be equal or better, is there a reason to keep the bugged version in the object? Then the version with the delta can be avoided.