@Sam-Harper 2019-04-10
Hi Jonas,  so no, I dont know exactly. Its completely unclear to me, its certainly something to study.  This also impacts the cut based ID in a slightly different way.  Time is a bit short, so  I would really like to keep the bugged variable around as its safer. Also there is the constraint that in 10_6 we need to be able to read 10_2 data and the IDs when re-run should give exactly the same value. And if we put in just the fixed value, we would have to put the bugged value in its place for 10_2 made samples so the same function would return different variables depending on input sample which to me violates the principle of least surprise and so should be avoided.

Btw, I've been wading through a lot of crusty e/gamma photon code due to this PR, it was a breath of fresh air when I got to your MVA framework. Also thanks for putting a debugging mode in it, it made it trivial for me to find the MVA was changing due to the conversion vtx prob being different.