@peruzzim 2019-04-10
> Also there is the constraint that in 10_6 we need to be able to read 10_2 data and the IDs when re-run should give exactly the same value. And if we put in just the fixed value, we would have to put the bugged value in its place for 10_2 made samples so the same function would return different variables depending on input sample which to me violates the principle of least surprise and so should be avoided.

I completely agree - it should behave as if the definition of the input variables (i.e. recipe to get them from MINIAOD) changes according to the version of the ID, not to the version used to produce the input sample or to run on top of it.