@Sam-Harper 2019-04-10
okay I hadnt properly removed the tau test. Its gone fully now