@Sam-Harper 2019-04-15
Okay first results of timing and diskspace in:
data samples: 
EGamma 2018 , 316058, LS 60,  1K events 
JetHT 2018: 322252, LS 80, 1K events

size on disk (bytes)

| sample | ref |  new |  change  |  
| ---| --- |---|---|
| Egamma AOD |  431865103  |  432187368 | +0.07% |
| Egamma miniAOD | 58896326 | 58861553 | -0.06% |
| JetHT AOD | 410829202 |  411015969 | 0.04% |
| JetHT miniAOD | 55616816 | 55586650 | -0.06% | 

From jetHT:
patElectrons_slimmedElectrons__reRECO. 3047.91 821.307
patPhotons_slimmedPhotons__reRECO. 1000.13 231.005
to 
patElectrons_slimmedElectrons__reRECO. 3079.88 820.549
patPhotons_slimmedPhotons__reRECO. 909.728 222.536

From e/gamma:
patElectrons_slimmedElectrons__reRECO. 7400.65 1901.78
patPhotons_slimmedPhotons__reRECO. 4785.06 1168.24
to
patElectrons_slimmedElectrons__reRECO. 7480.99 1903.83
patPhotons_slimmedPhotons__reRECO. 4356.82 1120.37

For e/gamma AOD:
recoPhotons_gedPhotons__reRECO. 2618.28 1885.62
recoGsfElectrons_gedGsfElectrons__reRECO. 2646 1785.37
to
recoPhotons_gedPhotons__reRECO. 2669.03 1984.21
recoGsfElectrons_gedGsfElectrons__reRECO. 2686.36 1808.91

This is as expected, we reduce the user data for pat::Electrons/Photons so miniAOD works out marginally smaller. AOD we only add info so its a bit bigger but AOD is so large w.r.t the e/gamma collections there is no impact.

Timing: E/gamma (RECO + MINIAOD job)

| delta/mean | delta/orJob | original | new | name  | 
| ---| --- |---|---|---|
| removed | -0.01% | 1.55 ms/ev | 0.00 ms/ev | heepIDVarValueMaps |
| removed | -0.05% | 8.23 ms/ev | 0.00 ms/ev | photonIDValueMapProducer |
| added | +0.09% | 0.00 ms/ev | 13.68 ms/ev | photonIDValueMapProducerRECO |
| removed | -0.00% | 0.25 ms/ev | 0.00 ms/ev | egmPhotonIsolation |
| -1.808031 | -0.01% | 0.97 ms/ev | 0.05 ms/ev | ctppsDiamondRecHits |
| -1.742736 | -0.02% | 3.32 ms/ev | 0.23 ms/ev | patJetCorrFactorsAK8 |
| -1.662091 | -0.01% | 0.87 ms/ev | 0.08 ms/ev | patJetCorrFactorsAK8Puppi |
| -1.635088 | -0.01% | 1.04 ms/ev | 0.10 ms/ev | gtDigis |
| +1.151515 | +0.00% | 0.01 ms/ev | 0.03 ms/ev | ak4PFCHSResidualCorrector |
| -1.121255 | -0.04% | 8.31 ms/ev | 2.34 ms/ev | dt1DRecHits |
| +0.804424 | +0.03% | 3.41 ms/ev | 7.99 ms/ev | muonCSCDigis |
| -0.723438 | -0.01% | 1.71 ms/ev | 0.80 ms/ev | muonDTDigis |
| -0.561243 | -0.00% | 1.40 ms/ev | 0.79 ms/ev | QGTagger |
| -0.557611 | -0.03% | 10.58 ms/ev | 5.96 ms/ev | csc2DRecHits |
| +0.417690 | +0.00% | 0.97 ms/ev | 1.48 ms/ev | castorreco |
| -0.401671 | -0.00% | 2.01 ms/ev | 1.34 ms/ev | MeasurementTrackerEventPreSplitting |
| +0.388164 | +0.03% | 9.89 ms/ev | 14.65 ms/ev | pfCombinedCvsLJetTags |
| +0.387707 | +0.00% | 1.36 ms/ev | 2.02 ms/ev | ecalRecHit |
| -0.376404 | -0.01% | 2.96 ms/ev | 2.02 ms/ev | gedGsfElectrons |
| -0.337486 | -0.00% | 1.03 ms/ev | 0.73 ms/ev | oniaPhotonCandidates |
| -0.290987 | -0.00% | 0.81 ms/ev | 0.60 ms/ev | photonMVAValueMapProducer |
| +0.288276 | +0.01% | 6.11 ms/ev | 8.17 ms/ev | pfCombinedCvsBJetTags |
| +0.281170 | +0.00% | 1.35 ms/ev | 1.79 ms/ev | dt1DCosmicRecHits |
| +0.247317 | +0.00% | 0.94 ms/ev | 1.20 ms/ev | totemRPUVPatternFinder |
| +0.215008 | +0.01% | 5.03 ms/ev | 6.24 ms/ev | pfImpactParameterTagInfosSlimmedDeepFlavour |
| -0.155100 | -0.01% | 10.45 ms/ev | 8.95 ms/ev | initialStepSeedsPreSplitting |
| +0.154680 | +0.03% | 30.14 ms/ev | 35.19 ms/ev | siPixelClustersPreSplitting |
| +0.144300 | +0.01% | 10.24 ms/ev | 11.84 ms/ev | gedGsfElectronsTmp |
| +0.137533 | +0.02% | 21.20 ms/ev | 24.33 ms/ev | lowPtGsfElectrons |
| +0.132523 | +0.00% | 4.92 ms/ev | 5.62 ms/ev | ecalDrivenGsfElectrons |
| +0.107962 | +0.01% | 7.68 ms/ev | 8.55 ms/ev | siStripZeroSuppression |
| +0.074234 | +0.01% | 11.58 ms/ev | 12.48 ms/ev | siPixelRecHitsPreSplitting |
| -0.066034 | -0.00% | 10.50 ms/ev | 9.83 ms/ev | ecalPreshowerRecHit |
| +0.060958 | +0.00% | 8.35 ms/ev | 8.88 ms/ev | siPixelDigis |

total 15.1542 s/ev  -> 15.3237 s/ev

JetHT (RECO + MINIAOD job)

| delta/mean | delta/orJob | original | new | name  | 
| ---| --- |---|---|---|
| removed | -0.00% | 0.84 ms/ev | 0.00 ms/ev | heepIDVarValueMaps |
| removed | -0.01% | 1.72 ms/ev | 0.00 ms/ev | photonIDValueMapProducer |
| added | +0.03% | 0.00 ms/ev | 4.31 ms/ev | photonIDValueMapProducerRECO |
| removed | -0.00% | 0.07 ms/ev | 0.00 ms/ev | egmPhotonIsolation |
| +1.971119 | +0.01% | 0.01 ms/ev | 1.10 ms/ev | totemRPUVPatternFinder |
| +1.900884 | +1.22% | 5.48 ms/ev | 215.60 ms/ev | csc2DRecHits |
| +1.884514 | +0.00% | 0.01 ms/ev | 0.37 ms/ev | hpsPFTauDiscriminationByMVA6LooseElectronRejectionBoosted |
| -1.875000 | -0.00% | 0.06 ms/ev | 0.00 ms/ev | ak4PFCHSL3AbsoluteCorrectorNoHF |
| -1.724179 | -0.07% | 12.53 ms/ev | 0.93 ms/ev | ctppsDiamondRawToDigi |
| +1.709538 | +1.15% | 16.74 ms/ev | 213.74 ms/ev | l1extraParticles |
| -1.573333 | -0.00% | 0.07 ms/ev | 0.01 ms/ev | ak4PFCHSResidualCorrector |
| -1.562842 | -0.00% | 0.33 ms/ev | 0.04 ms/ev | castorreco |
| -1.536504 | -0.01% | 1.32 ms/ev | 0.17 ms/ev | totemRPRawToDigi |
| +1.520327 | +0.54% | 14.72 ms/ev | 108.02 ms/ev | siStripZeroSuppression |
| +1.479872 | +0.55% | 16.68 ms/ev | 111.62 ms/ev | ecalPreshowerRecHit |
| -1.466667 | -0.00% | 0.01 ms/ev | 0.00 ms/ev | ak4PFPuppiL3AbsoluteCorrector |
| +1.455499 | +0.06% | 1.92 ms/ev | 12.19 ms/ev | ecalRecHit |
| -1.441860 | -0.00% | 0.07 ms/ev | 0.01 ms/ev | ak4PFCHSL2RelativeCorrectorNoHF |
| +1.434316 | +0.01% | 0.21 ms/ev | 1.28 ms/ev | displacedMuonSeeds |
| +1.108128 | +0.00% | 0.30 ms/ev | 1.04 ms/ev | cosmicsVetoTracksRaw |
| +1.062812 | +0.01% | 0.47 ms/ev | 1.54 ms/ev | ctppsPixelClusters |
| +0.976476 | +0.49% | 44.53 ms/ev | 129.50 ms/ev | siPixelClustersPreSplitting |
| +0.964688 | +0.01% | 0.69 ms/ev | 1.97 ms/ev | dt1DCosmicRecHits |
| -0.909688 | -0.06% | 15.95 ms/ev | 5.98 ms/ev | muonCSCDigis |
| -0.693323 | -0.01% | 1.68 ms/ev | 0.82 ms/ev | pfCombinedCvsBJetTagsPuppi |
| -0.647872 | -0.01% | 2.40 ms/ev | 1.22 ms/ev | ctppsPixelLocalTracks |
| +0.633344 | +0.57% | 105.69 ms/ev | 203.65 ms/ev | ecalMultiFitUncalibRecHit |
| -0.629874 | -0.00% | 1.10 ms/ev | 0.57 ms/ev | initialStepSeedLayersPreSplitting |
| +0.560400 | +0.01% | 2.30 ms/ev | 4.10 ms/ev | dt1DRecHits |
| -0.504625 | -0.05% | 19.63 ms/ev | 11.72 ms/ev | siPixelRecHitsPreSplitting |
| -0.497182 | -0.00% | 1.00 ms/ev | 0.60 ms/ev | oniaPhotonCandidates |
| -0.477415 | -0.01% | 3.95 ms/ev | 2.43 ms/ev | pfCombinedCvsLJetTagsPuppi |
| -0.459640 | -0.01% | 3.59 ms/ev | 2.25 ms/ev | MeasurementTrackerEventPreSplitting |
| -0.374311 | -0.00% | 2.48 ms/ev | 1.70 ms/ev | gedGsfElectrons |
| +0.295952 | +0.01% | 6.51 ms/ev | 8.78 ms/ev | pfCombinedCvsBJetTags |
| +0.279777 | +0.01% | 7.24 ms/ev | 9.60 ms/ev | standAloneMuons |
| +0.229283 | +0.02% | 11.87 ms/ev | 14.94 ms/ev | pfCombinedCvsLJetTags |
| +0.195334 | +0.01% | 7.79 ms/ev | 9.48 ms/ev | siPixelDigis |
| +0.138854 | +0.01% | 7.44 ms/ev | 8.55 ms/ev | gedGsfElectronsTmp |
| +0.133869 | +0.02% | 27.18 ms/ev | 31.08 ms/ev | lowPtGsfElectrons |
| +0.107189 | +0.00% | 6.21 ms/ev | 6.91 ms/ev | pfImpactParameterTagInfosSlimmedDeepFlavour |
| -0.096355 | -0.00% | 9.12 ms/ev | 8.28 ms/ev | initialStepSeedsPreSplitting |
| -0.082210 | -0.19% | 420.98 ms/ev | 387.74 ms/ev | tobTecStepTrackCandidates |
| +0.075587 | +0.02% | 35.86 ms/ev | 38.68 ms/ev | candidateVertexArbitratorCvsL |
| +0.068442 | +0.01% | 25.70 ms/ev | 27.52 ms/ev | candidateVertexArbitrator |
| -0.063223 | -0.10% | 291.32 ms/ev | 273.47 ms/ev | initialStepTrackCandidatesPreSplitting |
| -0.059326 | -0.26% | 783.14 ms/ev | 738.02 ms/ev | pixelLessStepTrackCandidates |
| +0.058450 | +0.01% | 33.90 ms/ev | 35.94 ms/ev | initialStepHitQuadrupletsPreSplitting |
| -0.057922 | -0.29% | 876.19 ms/ev | 826.87 ms/ev | jetCoreRegionalStepTrackCandidates |
| -0.056531 | -0.03% | 78.70 ms/ev | 74.38 ms/ev | detachedQuadStepTrackCandidates |
| -0.053005 | -0.16% | 534.54 ms/ev | 506.94 ms/ev | lowPtQuadStepTrackCandidates |
| -0.050711 | -0.15% | 532.72 ms/ev | 506.38 ms/ev | highPtTripletStepTrackCandidates |

event 17.1663 -> 17.6425 s/ev

so timing doesnt look that concerning and this is without the speedup that PhotonIDValueMapProducer already has in another PR.